refactor: split session runtime orchestration
